1)  Updated to latest upstream SVN trunk 2)  Fixed memleak in Alg_atoms::~Alg_atoms() 3)  Fixed memleak in Serial_buffer::~Serial_buffer() 4)  Added a couple more #pragmas to quiet MSVC a bit 5)  Changed Serial_buffer::init_for_read() to allocate a new buffer and copy the original into it.  I did this because Serial_buffer::check_buffer() can reallocate the buffer to make it bigger.  However, the caller of init_for_read() doesn't know this happened and when the original buffer passed in is no longer valid and causes a crash when freed by the caller. 6)  Fixed memleak in Alg_track::~Alg_track() 7)  Fixed memleak in Alg_seq::~Alg_seq()  (the Alg_track destructor wasn't being called) 8)  Changed stream parameter to Alg_seq::smf_write() and Alg_smf_write::write() from ofstream to ostream.  This should allow writing to any of the various streams. 9)  Changed a few casts to bool to generate proper value instead. 10) Fixed a bug in Alq_track::unserialize_track() that only appeared in release builds (when assert wasn't defined).  Two problems existed with the first being that the ALGT header was only bypassed in debug builds since the get_char()s were done in the assertions.  The second problem was the the header needed to be skipped in Alg_seq::unserialize_seq() since there's two paths into unserialize_track() and one path already stripped the header, while the other did not. 11) Fixed memleak in Alg_tracks::~Alg_tracks 12) Reverted one of the recent changes to allegrord.cpp that used a Windows specific function 13) Got rid of a couple more MSVC warnings
Patch sent to Roger...

+++ sf-cvs/trunk/lib-src/portsmf/allegrord.cpp
@@ -346,8 +346,8 @@ bool Alg_reader::parse()
                     note_ptr->time = time;
                     note_ptr->dur = dur;
                     note_ptr->set_identifier(key);
-                    note_ptr->pitch = pitch;
-                    note_ptr->loud = loud;
+                    note_ptr->pitch = (float) pitch;
+                    note_ptr->loud = (float) loud;
                     note_ptr->parameters = attributes;
                     seq->add_event(note_ptr, track_num); // sort later
                     if (seq->get_real_dur() < (time + dur)) seq->set_real_dur(time + dur);
@@ -678,6 +678,8 @@ bool Alg_reader::parse_val(Alg_parameter_ptr param, string &s, int i)
         // so total memory to allocate is (len - i) - 1
         char *r = new char[(len - i) - 1];
         strncpy(r, s.c_str() + i + 1, (len - i) - 2);
+//      Windows only
+//      strncpy_s(r, (len - i) - 2, s.c_str() + i + 1, _TRUNCATE);
         r[(len - i) - 2] = 0; // terminate the string
         param->s = r;
     } else if (s[i] == '\'') {
